Fixed bug present because of trying to save and restore the  array - it does not appear to be neccessary to explicitly do this as  seems to be dynamically managed by the server.
Modified the save and restore globals functions so that they will only try and restore the globals if they were freshly saved - this is to protect against a possible bug that occurs if a user reloads a page with a session id in the query string after the script successfully ran (and thus destroyed the session).

Modified the value stored in the test cookie to be the session id rather than 'true'.

Modified the way session ids are passed around - now just the random number portion is passed around, the stem is prepended to it every time within the script. Theoretically this provides a security benefit if the developer changes the stem from its default value to a secret value that only he is aware of.

Modified the cookie_cutter function to reload the page once more if cookies are enabled, this time with the original query string. In this way, for a user who has cookies enabled, then final result (including the address bar) looks exactly as it would have if CookieCheck wasn't used at all - the only difference being that the page reloaded itself a couple of times (though most of the time this is handled fairly transparently to the user anyway.

Created a global variable  that is used to store the error message if a cookie_cutter fails, rather than automatically printing the result out.

If testing of this revision goes ok, it will be released as v1.1.