Version 1.1 Release
Includes security fix for the default value of the session save path. Also fixes some minor bugs and makes some enhancements to the way the script affects the final user experience - at present (other than automatically reloading a couple of times) there is very little difference that the developer or end user will see whether the cookiecheck script is used or not.